+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64/FitCloudKit.framework/Headers/FitCloudIncomingCallPhotoSlotModel.h
@@ -1,0 +1,23 @@
+//
+//  FitCloudIncomingCallPhotoSlotModel.h
+//  FitCloudKit
+//
+//  Created by pcjbird on 8/2/25.
+//  Copyright © 2025 Zero Status. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface FitCloudIncomingCallPhotoSlotModel : NSObject
+
+@property (nonatomic, assign) NSUInteger slotIndex;
+
+@property (nonatomic, strong, nullable) NSString* phoneNumber;
+
+@property (nonatomic, readonly) BOOL hasPhoto;
+
+@end
+
+NS_ASSUME_NONNULL_END